fix(security): extract image-decode limits to standalone module to avoid model-load side effects
Pass-11 MM's `import { MAX_IMAGE_PIXELS } from '../file-upload.service'` in LocalStorageProvider
transitively pulled in User + FileUpload models via file-upload.service.ts, which triggered
env validation at module load. The existing `services/storage/init.test.ts` tests only mock
`../../../config` (not the model layer), so their `await import('../../../services/storage')`
now hit `DEV_DB_NAME is required` / `Missing UPLOAD_SIGNING_SECRET` errors and failed in CI.

Moved MAX_IMAGE_PIXELS + MAX_IMAGE_DIMENSION into a standalone constants module with no
service-layer imports. file-upload.service.ts and local-storage-provider.ts both import
from there; storage tests can now load the storage layer without triggering env validation.
